+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/VideoViewHolder.kt
@@ -16,6 +16,7 @@
 
 package im.vector.lib.attachmentviewer
 
+import android.util.Log
 import android.view.View
 import android.widget.ImageView
 import android.widget.ProgressBar
@@ -60,6 +61,13 @@ class VideoViewHolder constructor(itemView: View) :
 
     fun videoReady(file: File) {
         mVideoPath = file.path
+        if (isSelected) {
+            startPlaying()
+        }
+    }
+
+    fun videoReady(path: String) {
+        mVideoPath = path
         if (isSelected) {
             startPlaying()
         }
@@ -118,8 +126,13 @@ class VideoViewHolder constructor(itemView: View) :
                         eventListener?.get()?.onEvent(AttachmentEvents.VideoEvent(isPlaying, progress, duration))
                     }
         }
+        try {
+            videoView.setVideoPath(mVideoPath)
+        } catch (failure: Throwable) {
+            // Couldn't open
+            Log.v(VideoViewHolder::class.java.name, "Failed to start video")
+        }
 
-        videoView.setVideoPath(mVideoPath)
         if (!wasPaused) {
             videoView.start()
             if (progress > 0) {