+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/send/SendService.kt
@@ -110,13 +110,13 @@ interface SendService {
      * Schedule this message to be resent
      * @param localEcho the unsent local echo
      */
-    fun resendTextMessage(localEcho: TimelineEvent): Cancelable?
+    fun resendTextMessage(localEcho: TimelineEvent): Cancelable
 
     /**
      * Schedule this message to be resent
      * @param localEcho the unsent local echo
      */
-    fun resendMediaMessage(localEcho: TimelineEvent): Cancelable?
+    fun resendMediaMessage(localEcho: TimelineEvent): Cancelable
 
     /**
      * Remove this failed message from the timeline
@@ -124,7 +124,15 @@ interface SendService {
      */
     fun deleteFailedEcho(localEcho: TimelineEvent)
 
+    /**
+     * Delete all the events in one of the sending states
+     */
     fun clearSendingQueue()
+
+    /**
+     * Cancel sending a specific event. It has to be in one of the sending states
+     */
+    fun cancelSend(eventId: String)
 
     /**
      * Resend all failed messages one by one (and keep order)