+++ vector/src/main/java/im/vector/app/features/home/room/detail/timeline/helper/ContentUploadStateTrackerBinder.kt
@@ -75,9 +75,9 @@ private class ContentMediaProgressUpdater(private val progressLayout: ViewGroup,
             is ContentUploadStateTracker.State.Idle                -> handleIdle()
             is ContentUploadStateTracker.State.EncryptingThumbnail -> handleEncryptingThumbnail()
             is ContentUploadStateTracker.State.UploadingThumbnail  -> handleProgressThumbnail(state)
-            is ContentUploadStateTracker.State.Encrypting          -> handleEncrypting()
+            is ContentUploadStateTracker.State.Encrypting          -> handleEncrypting(state)
             is ContentUploadStateTracker.State.Uploading           -> handleProgress(state)
-            is ContentUploadStateTracker.State.Failure             -> handleFailure(state)
+            is ContentUploadStateTracker.State.Failure             -> handleFailure(/*state*/)
             is ContentUploadStateTracker.State.Success             -> handleSuccess()
         }
     }
@@ -98,26 +98,29 @@ private class ContentMediaProgressUpdater(private val progressLayout: ViewGroup,
     }
 
     private fun handleEncryptingThumbnail() {
-        doHandleEncrypting(R.string.send_file_step_encrypting_thumbnail)
+        doHandleEncrypting(R.string.send_file_step_encrypting_thumbnail, 0, 0)
     }
 
     private fun handleProgressThumbnail(state: ContentUploadStateTracker.State.UploadingThumbnail) {
         doHandleProgress(R.string.send_file_step_sending_thumbnail, state.current, state.total)
     }
 
-    private fun handleEncrypting() {
-        doHandleEncrypting(R.string.send_file_step_encrypting_file)
+    private fun handleEncrypting(state: ContentUploadStateTracker.State.Encrypting) {
+        doHandleEncrypting(R.string.send_file_step_encrypting_file, state.current, state.total)
     }
 
     private fun handleProgress(state: ContentUploadStateTracker.State.Uploading) {
         doHandleProgress(R.string.send_file_step_sending_file, state.current, state.total)
     }
 
-    private fun doHandleEncrypting(resId: Int) {
+    private fun doHandleEncrypting(resId: Int, current: Long, total: Long) {
         progressLayout.visibility = View.VISIBLE
+        val percent = if (total > 0) (100L * (current.toFloat() / total.toFloat())) else 0f
         val progressBar = progressLayout.findViewById<ProgressBar>(R.id.mediaProgressBar)
         val progressTextView = progressLayout.findViewById<TextView>(R.id.mediaProgressTextView)
-        progressBar?.isIndeterminate = true
+        progressBar?.isIndeterminate = false
+        progressBar?.progress = percent.toInt()
+        progressTextView.isVisible = true
         progressTextView?.text = progressLayout.context.getString(resId)
         progressTextView?.setTextColor(messageColorProvider.getMessageTextColor(SendState.ENCRYPTING))
     }
@@ -130,19 +133,23 @@ private class ContentMediaProgressUpdater(private val progressLayout: ViewGroup,
         progressBar?.isVisible = true
         progressBar?.isIndeterminate = false
         progressBar?.progress = percent.toInt()
+        progressTextView.isVisible = true
         progressTextView?.text = progressLayout.context.getString(resId,
                 TextUtils.formatFileSize(progressLayout.context, current, true),
                 TextUtils.formatFileSize(progressLayout.context, total, true))
         progressTextView?.setTextColor(messageColorProvider.getMessageTextColor(SendState.SENDING))
     }
 
-    private fun handleFailure(state: ContentUploadStateTracker.State.Failure) {
+    private fun handleFailure(/*state: ContentUploadStateTracker.State.Failure*/) {
         progressLayout.visibility = View.VISIBLE
         val progressBar = progressLayout.findViewById<ProgressBar>(R.id.mediaProgressBar)
         val progressTextView = progressLayout.findViewById<TextView>(R.id.mediaProgressTextView)
         progressBar?.isVisible = false
-        progressTextView?.text = errorFormatter.toHumanReadable(state.throwable)
-        progressTextView?.setTextColor(messageColorProvider.getMessageTextColor(SendState.UNDELIVERED))
+        // Do not show the message it's too technical for users, and unfortunate when upload is cancelled
+        // in the middle by turning airplane mode for example
+        progressTextView.isVisible = false
+//        progressTextView?.text = errorFormatter.toHumanReadable(state.throwable)
+//        progressTextView?.setTextColor(messageColorProvider.getMessageTextColor(SendState.UNDELIVERED))
     }
 
     private fun handleSuccess() {